finalized config.ini

+++ src/fileManager.py
@@ -23,7 +23,6 @@ class FileManager:
         self.verifyDirectories(proj)
         
     def verifyDirectories(self, proj):
-        print(proj)
 
         if '/' in proj:
             proj = proj.replace('/', '+')
@@ -57,16 +56,19 @@ class FileManager:
         self.f_dir = self.f_dir + proj
 
     def verifyConfig(self, repo, proj, depth, f_dir):
-        config = configparser.ConfigParser()
-        config.read('config.ini')
-        config.add_section('Operation Atributes')
-        config.set('Operation Atributes', 'repository', repo)
-        config.set('Operation Atributes', 'project link', proj)
-        config.set('Operation Atributes', 'maximum depth', depth)
-        config.set('Operation Atributes', 'end directory', f_dir)
 
-        with open(self.p_dir + '/config.ini', 'w', newline='') as writeConfig:
-            config.write(writeConfig)
+        if not os.path.exists('config.ini'):
+            config = configparser.ConfigParser()
+            config.read('config.ini')
+            config.add_section('Operation Atributes')
+            config.set('Operation Atributes', 'repository', repo)
+            config.set('Operation Atributes', 'project link', proj)
+            config.set('Operation Atributes', 'maximum depth', depth)
+            config.set('Operation Atributes', 'end directory', f_dir)
+
+            with open(self.p_dir + '/config.ini', 'w', newline='') as writeConfig:
+                config.write(writeConfig)
+            writeConfig.close()
 
     def addLinks(self, module, dependencies):
